Test the scene manager
