+++ TestSceneManage/Scene.java
@@ -1,0 +1,8 @@
+import java.awt.Graphics;
+
+public interface Scene {
+    void init();
+    void update();
+    void render(Graphics g);
+    void dispose();
+}
